# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PovarUp Marketplace API
  version: 1.0.0
servers:
  - url: http://localhost:4000/api
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but not authorized for this resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
    Shift:
      type: object
      properties:
        id: { type: string }
        businessId: { type: string }
        locationId: { type: string }
        title: { type: string }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        payRateCents: { type: integer }
        status: { type: string }
...
paths:
...
  /auth/session:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { userId: { type: string }, role: { type: string } }
      responses: { '200': { description: session } }
...
  /shifts:
    get:
      responses:
        '200': { description: list shifts, content: { application/json: { schema: { type: object } } } }
    post:
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Shift' } } } }
      responses: { '201': { description: created } }
  /applications:
    get:
      description: |
        Returns only applications relevant to the authenticated actor (worker owns application,
        business owns related shift) unless role is admin.
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: list apps }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { shiftId: { type: string }, workerId: { type: string } }
      responses: { '201': { description: created } }
  /assignments:
    get:
      description: Returns assignment list filtered to caller ownership; non-owners receive an empty list.
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: list assignments }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /assignments/offer:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, properties: { applicationId: { type: string }, businessId: { type: string } } }
      responses: { '201': { description: offered } }
...